build(pojo): move some class to pojo package
