+++ pom.xml
@@ -17,7 +17,7 @@
     <properties>
         <java.version>1.8</java.version>
         <mysql.connector.java.version>5.1.47</mysql.connector.java.version>
-        <mybatis.spring.boot.starter.version>2.1.4</mybatis.spring.boot.starter.version>
+        <mybatis.spring.boot.starter.version>2.2.0</mybatis.spring.boot.starter.version>
         <druid.spring.boot.starter.version>1.1.10</druid.spring.boot.starter.version>
         <pagehelper.spring.boot.starter.version>1.4.7</pagehelper.spring.boot.starter.version>
         <lombok.version>1.18.18</lombok.version>
